fix float16 precomputed code issues (#1406)
Summary:
Pull Request resolved: https://github.com/facebookresearch/faiss/pull/1406

I appear to have broken this with the rework of float16 support in Faiss GPU, though I cannot figure out why the tests only started failing recently.

cuBLAS does not support a f16 x f32 = f32 matrix multiplication. With a f16 coarse quantizer and IVFPQ precomputed codes, we were attempting to perform such a multiplication.

Now, in the precomputed code calculation, we intercept this and change it to a f32 x f32 = f32 computation.

The test when run by itself was also failing separately, though when run in series with the other test_gpu_index_ivfpq tests it was succeeding, due to the fact that the seed is only initialized once. The epsilon needed to change a slight bit.

Reviewed By: mdouze

Differential Revision: D23687070

fbshipit-source-id: 14a535407ed433eeaef3bc77cb0d6f5909c55b9f