Nightly failure fix - ignore searched vectors with identical distances (#3809)
Summary:
Pull Request resolved: https://github.com/facebookresearch/faiss/pull/3809

test_IndexIVFRQ is failing nightly GPU w/RAFT builds. These are false positive failures for the edge cases when the distance of a query vector to two candidate vectors is identical. In such cases, the search and search_and_reconstruct return these candidate vectors in a different order (on GPU w/Raft).

Current fix is to ignore the order of such ids with identical distances by sorting on distance and use id as tiebreaker. There is negligible performance implication (5ms extra), given the test candidate size is low.

Reviewed By: mengdilin, ramilbakhshyiev

Differential Revision: D61977915

fbshipit-source-id: 5f7a81c51c91a967013fb5e69e2bfee59be341c7