Allow Odd Degree Binary CAGRA graph to HNSW conversion (#4516)
Summary:
The check on the graph degree to ensure that the Binary CAGRA graph is convertible to HNSW is too strict. Instead, an odd degree graph can be converted to HNSW just as well by dropping the farthest edge for each vertex. This is also consistent with regular CAGRA to HNSW conversion.

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/4516

Reviewed By: junjieqi

Differential Revision: D79837285

Pulled By: mnorris11

fbshipit-source-id: 0dca1efa3fbe9d668c0896f1287ed51291b89b01